server/pyproject.toml
[project]
name = "praxisdoktor-server"
version = "0.2.0"
description = "PraxisDoktor backend — FastAPI + Postgres skeleton"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14",
    "asyncpg>=0.30",
    "psycopg[binary]>=3.2",
    "argon2-cffi>=23.1",
    "itsdangerous>=2.2",
    "pydantic-settings>=2.6",
    "python-multipart>=0.0.20",
    "aiofiles>=24.1",
    "websockets>=14",
    # intake (ported from v1)
    "faster-whisper>=1.0",
    "easyocr>=1.7",
    "soundfile>=0.12",
    "av>=13",
    "scipy>=1.14",
    "numpy>=1.26,<2",
    "speechbrain>=1.0",
    "torch>=2.4",
    "torchaudio>=2.4",
    "scikit-learn>=1.5",
    "httpx>=0.27",
]

[tool.uv]
package = false
